+++ Libraries/DeviceList/DeviceList/Coordinators/DeviceListCoordinatorImpl.swift
@@ -111,7 +111,7 @@ extension DeviceListCoordinatorImpl: SettingDelegate {
         let session = scope.getService(SessionAPI.self)
         session.logout()
         let loginAPI = scope.getService(LoginAPI.self)
-        let navigator = SHNavigationView(scope: scope) { [weak self] router -> AnyView in
+        let navigator = OMENavigationView(scope: scope) { [weak self] router -> AnyView in
             self?.login = loginAPI.createLoginCoordinator(router: router)
             return self?.login?.createLoginView() ?? AnyView(Text("Error"))
         }
