+++ OmniedgeNew/OmniedgeNew/Coordinator/AppCoordinator.swift
@@ -53,7 +53,7 @@ class AppCoordinator: Coordinator {
 
     private var loginView: AnyView {
         let loginAPI = scope.getService(LoginAPI.self)
-        let navigator = SHNavigationView(scope: scope) { [weak self] router -> AnyView in
+        let navigator = OMENavigationView(scope: scope) { [weak self] router -> AnyView in
             let coordinator = loginAPI.createLoginCoordinator(router: router)
             self?.child.append(coordinator)
             return coordinator.createLoginView()
@@ -63,7 +63,7 @@ class AppCoordinator: Coordinator {
 
     private func homeView(_ user: User, token: String) -> AnyView {
         let deviceList = scope.getService(DeviceListAPI.self)
-        let navigator = SHNavigationView(scope: scope) { [weak self] router -> AnyView in
+        let navigator = OMENavigationView(scope: scope) { [weak self] router -> AnyView in
             let coordinator = deviceList.createHomeCoordinator(router: router, user: user, token: token)
             self?.child.append(coordinator)
             return coordinator.createHomePage()
